+++ tests/basic.test.tsx
@@ -7,7 +7,7 @@ import * as React from "react";
 import { MemoryRouter } from "react-router-dom";
 import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
 import { Providers } from "../src/app/providers";
-import PostsManager from "../src/pages/PostsManagerPage";
+import PostsManager from "../src/pages/PostsManager";
 import { TEST_POSTS, TEST_SEARCH_POST, TEST_USERS } from "./mockData";
 // MSW 서버 설정
 const server = setupServer(
